refactor: remove governance content component
